+++ src/textual/scrollbar.py
@@ -286,6 +286,9 @@ class ScrollBar(Widget):
         else:
             background = styles.scrollbar_background
             color = styles.scrollbar_color
+        if background.a < 1:
+            base_background, _ = self.parent._opacity_background_colors
+            background = base_background + background
         color = background + color
         scrollbar_style = Style.from_color(color.rich_color, background.rich_color)
         if self.screen.styles.scrollbar_color.a == 0: